sms/requirements.txt
streamlit>=1.33
pandas>=2.1
PyYAML>=6.0
requests>=2.31
python-dateutil>=2.8
streamlit-calendar>=1.3.1
